adds red tests for auth registration
