+++ tests/rest_api/test_custom_auth.py
@@ -1,0 +1,69 @@
+from base64 import b64encode
+import pytest
+from typing import Any, cast, Dict
+from sources import rest_api
+from sources.rest_api.typing import ApiKeyAuthConfig, AuthConfig
+from dlt.sources.helpers.rest_client.auth import APIKeyAuth, OAuth2ClientCredentials
+
+class CustomOAuth2(OAuth2ClientCredentials):
+    def build_access_token_request(self) -> Dict[str, Any]:
+        """Used e.g. by Zoom Zoom Video Communications, Inc."""
+        authentication: str = b64encode(
+            f"{self.client_id}:{self.client_secret}".encode()
+        ).decode()
+        return {
+            "headers": {
+                "Authorization": f"Basic {authentication}",
+                "Content-Type": "application/x-www-form-urlencoded",
+            },
+            "data": self.access_token_request_data,
+        }
+
+class TestCustomAuth:
+    @pytest.fixture
+    def custom_auth_config(self) -> AuthConfig:
+        config: AuthConfig = {
+            "type": "custom_oauth_2",  # type: ignore
+            "access_token_url": "https://example.com/oauth/token",
+            "client_id": "test_client_id",
+            "client_secret": "test_client_secret",
+            "access_token_request_data": {
+                "grant_type": "account_credentials",
+                "account_id": "test_account_id",
+            },
+        }
+        return config
+
+    def test_creates_builtin_auth_without_registering(self) -> None:
+        config: ApiKeyAuthConfig = {
+            "type": "api_key",
+            "api_key": "test-secret",
+            "location": "header",
+        }
+        auth = cast(APIKeyAuth, rest_api.config_setup.create_auth(config))
+        assert auth.api_key == "test-secret"
+
+    def test_not_registering_throws_error(self, custom_auth_config: AuthConfig) -> None:
+        with pytest.raises(ValueError) as e:
+            rest_api.config_setup.create_auth(custom_auth_config)
+
+        assert e.match("Invalid authentication: custom_oauth_2.")
+
+    def test_registering_adds_to_AUTH_MAP(self, custom_auth_config: AuthConfig) -> None:
+        rest_api.config_setup.register_auth("custom_oauth_2", CustomOAuth2)
+        cls = rest_api.config_setup.get_auth_class("custom_oauth_2")
+        assert cls is CustomOauth2
+
+
+    def test_registering_adds_to_AUTH_MAP(self, custom_auth_config: AuthConfig) -> None:
+        rest_api.config_setup.register_auth("custom_oauth_2", CustomOAuth2)
+        auth = rest_api.config_setup.create_auth(custom_auth_config)
+        assert auth.build_access_token_request()["data"]["account_id"] == "test_account_id"
+
+    def test_registering_not_auth_config_base_throws_error(self):
+        class NotAuthConfigBase:
+            pass
+
+        with pytest.raises(ValueError) as e:
+            rest_api.config_setup.register_auth("not_an_auth_config_base", NotAuthConfigBase)
+        assert e.match("Invalid authentication: NotAuthConfigBase.")
